+++ workspace/plugins/monaco-editor/src/breadcrumb/asset.ts
@@ -82,6 +82,9 @@ export class AssetBreadcrumb extends Breadcrumb {
             });
         });
 
+        /* 当前文件对应的面包屑显示为激活状态 */
+        Breadcrumb.setLastBreadcrumbItemActive(breadcrumb);
+
         if (options.stores) {
             /* 全屏 */
             breadcrumb.breadcrumbIcons.push({